src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.2.3            # via django
attrs==19.3.0             # via jsonschema, pytest
certifi==2019.11.28       # via requests, sentry-sdk
chardet==3.0.4            # via requests
coreapi==2.3.3            # via drf-yasg
coreschema==0.0.4         # via coreapi, drf-yasg
coverage==5.0.3           # via pytest-cov
defusedxml==0.6.0         # via djangorestframework-xml
django-cors-headers==3.2.1
django-environ==0.4.5
django-filter==2.2.0
django-healthchecks==1.4.2
django-postgres-unlimited-varchar==1.1.0
django==3.0.2
djangorestframework-csv==2.1.0  # via drf-amsterdam
djangorestframework-gis==0.15
djangorestframework-xml==1.4.0  # via drf-amsterdam
djangorestframework==3.11.0
drf-amsterdam==0.3.2
drf-extensions==0.5.0     # via drf-amsterdam
drf-yasg==1.17.0
entrypoints==0.3          # via flake8
flake8==3.7.9
idna==2.8                 # via requests
inflection==0.3.1         # via drf-yasg
itypes==1.1.0             # via coreapi
jinja2==2.10.3            # via coreschema
jsonschema==3.2.0
markupsafe==1.1.1         # via jinja2
mccabe==0.6.1             # via flake8
more-itertools==8.1.0     # via pytest
packaging==20.0           # via drf-yasg, pytest
pluggy==0.13.1            # via pytest
psycopg2-binary==2.8.4
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8
pyparsing==2.4.6          # via packaging
pyrsistent==0.15.7        # via jsonschema
pytest-cov==2.8.1
pytest-django==3.8.0
pytest==5.3.4
python-string-utils==0.6.0
pytz==2019.3              # via django
readerwriterlock==1.0.6
requests==2.22.0
ruamel.yaml==0.16.6       # via drf-yasg
sentry-sdk==0.14.0
six==1.14.0               # via django-healthchecks, djangorestframework-csv, drf-yasg, jsonschema, packaging, pyrsistent
sqlparse==0.3.0           # via django
typing-extensions==3.7.4.1  # via readerwriterlock
unicodecsv==0.14.1        # via djangorestframework-csv
uritemplate==3.0.1        # via coreapi, drf-yasg
urllib3==1.25.7           # via requests, sentry-sdk
wcwidth==0.1.8            # via pytest

# The following packages are considered to be unsafe in a requirements file:
# setuptools
